\name{boral-package}

\alias{boral-package}

\docType{package}

\title{Bayesian Ordination and Regression AnaLysis (boral)}

\description{
\packageDescription{boral}
}

\details{
\tabular{ll}{
Package: \tab boral\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2014-12-12\cr
License: \tab GPL-2\cr
}
}

\author{
\packageAuthor{boral}

Maintainer: \packageMaintainer{boral}
}

\references{
\itemize{
\item Hui et al. (2014). Model-based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6, 399-411.

\item Plummer, M. (2003). JAGS: A program for analysis of Bayesian graphical models using Gibbs sampling. In Proceedings of the 3rd International Workshop on Distributed Statistical Computing. March (pp. 20-22).

\item Skrondal, A., and Rabe-Hesketh, S. (2004). Generalized latent variable modeling: Multilevel, longitudinal, and structural equation models. CRC Press.

\item Warton et al. (2015). So Many Variables: Joint Modeling in Community Ecology. Trends in Ecology and Evolution, 30, 766-779.

\item Yi W. et al. (2013). \code{mvabund}: statistical methods for analysing multivariate abundance data. R package version 3.8.4. } 
}

\examples{
## Please see main boral function for examples. 
}